api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.43.0
pdfplumber==0.11.4
requests==2.32.3
urllib3==2.2.3
python-multipart==0.0.9